better interact

+++ scripts/devel-tools.ps1
@@ -763,9 +763,11 @@ function remove_installation {
     winget uninstall --name "Windows Subsystem For Linux Update" | Out-Null
     winget uninstall --name "Windows Subsystem For Linux WSLg Preview" | Out-Null
     unset_dvlp_envs
-    Write-Host "`r`n`r`n"
-    Write-Host -nonewline "delete directory $env:USERPROFILE/repos/$($git_owner) ?"
-    Remove-Item "$env:USERPROFILE/repos/$($git_owner)" -Recurse -Confirm:$true -Force -ErrorAction SilentlyContinue
+    if (Test-Path "$env:USERPROFILE/repos/$($git_owner)") {
+        Write-Host "`r`n`r`n"
+        Write-Host -nonewline "delete directory $env:USERPROFILE/repos/$($git_owner) ?"
+        Remove-Item "$env:USERPROFILE/repos/$($git_owner)" -Recurse -Confirm:$true -Force -ErrorAction SilentlyContinue
+    }
     Remove-Item "$env:USERPROFILE/repos/$($git_owner)/.dvlp-installed" -Confirm:$false -Force -ErrorAction SilentlyContinue
     Remove-Item "$env:USERPROFILE/repos/$($git_owner)/.docker-installed" -Confirm:$false -Force -ErrorAction SilentlyContinue
     Remove-Item "$env:USERPROFILE/repos/$($git_owner)/.vscode-installed" -Confirm:$false -Force -ErrorAction SilentlyContinue
@@ -775,12 +777,17 @@ function remove_installation {
     Remove-Item "$env:USERPROFILE/repos/$($git_owner)/.hypervm-installed" -Confirm:$false -Force -ErrorAction SilentlyContinue
     Remove-Item "$env:USERPROFILE/repos/$($git_owner)/.python-installed" -Confirm:$false -Force -ErrorAction SilentlyContinue
     Remove-Item "$env:USERPROFILE/repos/$($git_owner)/.github-installed" -Confirm:$false -Force -ErrorAction SilentlyContinue
-    Write-Host "`r`n`r`n"
-    Write-Host "delete directory $env:USERPROFILE/kache ?"
-    Remove-Item "$env:USERPROFILE/kache" -Recurse -Confirm:$true -Force -ErrorAction SilentlyContinue
-    Write-Host "`r`n`r`n"
-    Write-Host "delete $env:USERPROFILE/.wslconfig ?"
-    Remove-Item "$env:USERPROFILE/.wslconfig" -Confirm:$true -Force -ErrorAction SilentlyContinue
+    if (Test-Path "$env:USERPROFILE/kache") {
+        Write-Host "`r`n`r`n"
+        Write-Host "delete directory $env:USERPROFILE/kache ?"
+        Remove-Item "$env:USERPROFILE/kache" -Recurse -Confirm:$true -Force -ErrorAction SilentlyContinue
+    }
+    if (Test-Path "$env:USERPROFILE/.wslconfig" -PathType Leaf){
+        Write-Host "`r`n`r`n"
+        Write-Host "delete $env:USERPROFILE/.wslconfig ?"
+        Remove-Item "$env:USERPROFILE/.wslconfig" -Confirm:$true -Force -ErrorAction SilentlyContinue
+    }
+
     # }
 
 }
